# [netcore] Add intrinsic implementation for Unsafe.InitBlockUnaligned