Fixes #16591