@filipnavara 2019-08-31
The "Linux x64 Bitcode+Interpreter" machine configuration is busted (/cc @directhex @akoeplinger).