@filipnavara 2019-08-31
While the benchmark results I got very not completely conclusive (some invocations of `ReadOnlySpan.IndexOfString` were slower to the factor of 1.5x to 1.9x; have to figure out why) the overall performance improved for the benchmarks I was targetting:

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Memory.Span<Byte>.Clear(Size: 512)                                        |     22.02 |           172.18 |             7.82 |         |
| System.Memory.Span<Byte>.Fill(Size: 512)                                         |     21.86 |           174.53 |             7.99 |         |
| System.Memory.Span<Char>.ToArray(Size: 512)                                      |      9.01 |           613.85 |            68.11 |         |
| System.Memory.Span<Int32>.ToArray(Size: 512)                                     |      8.26 |          1153.68 |           139.61 | several?|
| System.Memory.Memory<Char>.ToArray(Size: 512)                                    |      8.16 |           611.23 |            74.92 | bimodal |
| System.Memory.ReadOnlyMemory<Byte>.ToArray(Size: 512)                            |      7.60 |           318.00 |            41.84 | several?|
| System.Memory.Span<Byte>.ToArray(Size: 512)                                      |      7.10 |           312.93 |            44.06 | several?|
| System.Memory.ReadOnlyMemory<Char>.ToArray(Size: 512)                            |      6.72 |           623.69 |            92.83 | bimodal |
| System.Memory.Memory<Byte>.ToArray(Size: 512)                                    |      5.38 |           309.06 |            57.40 | bimodal |